refactor: :recycle: move internal code into `internal.py`
