--- src/seedcase_sprout/check_datapackage/add_package_recommendations.py
@@ -1,24 +1,0 @@
-from seedcase_sprout.check_datapackage.constants import (
-    NAME_PATTERN,
-    PACKAGE_RECOMMENDED_FIELDS,
-    SEMVER_PATTERN,
-)
-
-
-def add_package_recommendations(schema: dict) -> dict:
-    """Add recommendations from the Data Package standard to the schema.
-
-    Modifies the schema in place.
-
-    Args:
-        schema: The full Data Package schema.
-
-    Returns:
-        The updated Data Package schema.
-    """
-    schema["required"].extend(PACKAGE_RECOMMENDED_FIELDS.keys())
-    schema["properties"]["name"]["pattern"] = NAME_PATTERN
-    schema["properties"]["version"]["pattern"] = SEMVER_PATTERN
-    schema["properties"]["contributors"]["items"]["required"] = ["title"]
-    schema["properties"]["sources"]["items"]["required"] = ["title"]
-    return schema
