--- src/seedcase_sprout/check_datapackage/add_resource_recommendations.py
@@ -1,18 +1,0 @@
-from seedcase_sprout.check_datapackage.constants import NAME_PATTERN
-
-
-def add_resource_recommendations(schema: dict) -> dict:
-    """Add recommendations from the Data Resource standard to the schema.
-
-    Modifies the schema in place.
-
-    Args:
-        schema: The full Data Package schema.
-
-    Returns:
-        The updated Data Package schema.
-    """
-    schema["properties"]["resources"]["items"]["properties"]["name"]["pattern"] = (
-        NAME_PATTERN
-    )
-    return schema
