+++ src/seedcase_sprout/check_datapackage/check_properties.py
@@ -1,15 +1,11 @@
-from seedcase_sprout.check_datapackage.add_package_recommendations import (
-    add_package_recommendations,
-)
-from seedcase_sprout.check_datapackage.add_resource_recommendations import (
-    add_resource_recommendations,
-)
 from seedcase_sprout.check_datapackage.check_error import CheckError
-from seedcase_sprout.check_datapackage.check_object_against_json_schema import (
-    check_object_against_json_schema,
-)
 from seedcase_sprout.check_datapackage.constants import DATA_PACKAGE_SCHEMA_PATH
-from seedcase_sprout.internals import _read_json
+from seedcase_sprout.check_datapackage.internals import (
+    _add_package_recommendations,
+    _add_resource_recommendations,
+    _check_object_against_json_schema,
+    _read_json,
+)
 
 
 def check_properties(
@@ -36,7 +32,7 @@ def check_properties(
     schema = _read_json(DATA_PACKAGE_SCHEMA_PATH)
 
     if check_recommendations:
-        add_package_recommendations(schema)
-        add_resource_recommendations(schema)
+        _add_package_recommendations(schema)
+        _add_resource_recommendations(schema)
 
-    return check_object_against_json_schema(properties, schema)
+    return _check_object_against_json_schema(properties, schema)
