--- src/seedcase_sprout/check_datapackage/get_full_json_path_from_error.py
@@ -1,21 +1,0 @@
-import re
-
-from jsonschema import ValidationError
-
-
-def get_full_json_path_from_error(error: ValidationError) -> str:
-    """Returns the full `json_path` to the error.
-
-    For 'required' errors, the field name is extracted from the error message.
-
-    Args:
-        error: The error to get the full `json_path` for.
-
-    Returns:
-        The full`json_path` of the error.
-    """
-    if error.validator == "required":
-        match = re.search("'(.*)' is a required property", error.message)
-        if match:
-            return f"{error.json_path}.{match.group(1)}"
-    return error.json_path
